refactor(nextly): decide document visibility once, for both surfaces
The activity feed and the pending-edit cards ask the same question — which of
these content-keyed rows may this caller be told about — and each carried its
own answer. The copies did not stay equal. Per-locale grouping, the
registry-kind check, the stale-locale drop, bounded concurrency and resolving a
Single's live id before probing it were all present on the pending-edit side and
absent from the other, and every one of them is a defect on the side that lacks
it: a Single probed without resolving its live id first reads through a path
that AUTO-CREATES a missing Single, so loading a dashboard performs a write.

`services/lib/document-visibility.ts` is now that single decision, and both
surfaces are translations into it. `pending-edit-visibility.ts` keeps only what
it alone knows — that `VersionScopeKind` is wider than the two kinds a content
read path can answer for, so a `page`-scoped row resolves to nothing and is
dropped rather than sent to a service that does not hold the document.

It is separate from `readable-documents.ts` because the two answer different
questions: that module asks the read path which ids of ONE collection in ONE
language survive its rules, and this one decides which collections and languages
to ask about at all. Keeping them apart also keeps the lower one mockable from
the upper one's tests.

The fail-closed rule is written once. A rejected read has told us nothing, and
nothing must not read as visible; the two kinds needed that separately, and a
copy of it is a second place for the direction to drift.

The activity feed resolves the registry and locale snapshot ONCE per refill
rather than per page, for the reason the pending-edit walk does: both helpers
answer an unreachable dependency with an empty result, so a per-page read lets a
transient failure drop one page's rows while its neighbours keep theirs, and the
paging then reports a short feed as though it were whole.
